Move two variables to common place
