+++ tests/common/snappi_tests/traffic_generation.py
@@ -11,7 +11,7 @@ from tests.common.snappi_tests.common_helpers import get_egress_queue_count, pfc
     traffic_flow_mode
 from tests.common.snappi_tests.port import select_ports, select_tx_port
 from tests.common.snappi_tests.snappi_helpers import wait_for_arp, fetch_snappi_flow_metrics
-from tests.snappi_tests.variables import pfcQueueGroupSize, pfcQueueValueDict
+from variables import pfcQueueGroupSize, pfcQueueValueDict
 from tests.common.cisco_data import is_cisco_device
 
 logger = logging.getLogger(__name__)
