+++ tests/common/snappi_tests/variables.py
@@ -1,0 +1,9 @@
+pfcQueueGroupSize = 8  # can have values 4 or 8
+pfcQueueValueDict = {0: 0,
+                     1: 1,
+                     2: 0,
+                     3: 3,
+                     4: 2,
+                     5: 0,
+                     6: 1,
+                     7: 0}
