+++ tests/snappi_tests/variables.py
@@ -84,16 +84,6 @@ dut_ipv6_start = '2000:1::1'
 snappi_ipv6_start = '2000:1::2'
 v6_prefix_length = 126
 
-pfcQueueGroupSize = 8  # can have values 4 or 8
-pfcQueueValueDict = {0: 0,
-                     1: 1,
-                     2: 0,
-                     3: 3,
-                     4: 2,
-                     5: 0,
-                     6: 1,
-                     7: 0}
-
 
 def create_ip_list(value, count, mask=32, incr=0):
     '''
